backend/requirements.txt
annotated-types==0.7.0
anyio==4.9.0
asgiref==3.8.1
click==8.1.8
Django==5.1.4
django-environ==0.11.2
fastapi==0.115.12
h11==0.14.0
idna==3.10
iniconfig==2.0.0
mock==5.1.0
mysql-connector-python==9.1.0
mysqlclient==2.2.6
mysqlx-connector-python==9.1.0
packaging==24.2
pluggy==1.5.0
protobuf==4.25.3
psycopg2-binary==2.9.10
pydantic==2.11.3
pydantic_core==2.33.1
pytest==7.4.0
pytest-django==4.5.2
python-decouple==3.8
sniffio==1.3.1
sqlparse==0.5.3
starlette==0.46.2
typing-inspection==0.4.0
typing_extensions==4.13.2
uvicorn==0.34.1
